% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootCRCumInc.R
\name{bootCRCumInc}
\alias{bootCRCumInc}
\title{Bootstrap for estimated CIF}
\usage{
bootCRCumInc(df, exit, event, exposure, entry = NULL, weights = NULL,
  ipwvars = NULL, rep = 0, seed = 54321, print.attr = T)
}
\arguments{
\item{df}{A number.}

\item{exit}{A number.}

\item{event}{A number.}

\item{exposure}{A number.}

\item{entry}{A number.}

\item{weights}{A number.}

\item{ipwvars}{A number.}

\item{rep}{A number.}

\item{seed}{A number.}

\item{print.attr}{A number.}
}
\value{
Estimating CIF per event and exposure level
}
\description{
This function is based on the CRCumInc from this XX package.
Returns an object type dataframe with the following variables and order:
col1   col2    col3  col4
ttF1.lower  ttF1.upper  ttF2.lower  ttF2.upper
}
