% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper Function.R
\name{CRHaz_dx}
\alias{CRHaz_dx}
\title{Wrapper Function}
\usage{
CRHaz_dx(df, exit, event, exposure, entry = NULL, weights = NULL,
  ipwvars = NULL, maxtime = Inf, rep = NULL, eoi = -1)
}
\value{
CRHaz_dx
}
\description{
Main function
See help for other functions: check.packages, checking_data, data_CIF, fig_cif
}
