% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRhaz_dx.R
\docType{package}
\name{CRhazdx}
\alias{CRhazdx}
\alias{CRhazdx-package}
\title{CRhaz_dx: A non-Parametric Diagnostic Tool for Competing Risk hazard estimation}
\description{
The CRhaz_dx package provides a non-Parametric Diagnostic Tool for Competing Risk hazard estimation.
}
\section{competdiag functions}{

The CRhaz_dx functions ...
}

