% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRhaz_dx.R
\docType{package}
\name{hrcompetdx}
\alias{hrcompetdx}
\alias{hrcompetdx-package}
\title{hrcompetdx: A non-Parametric Diagnostic Tool for Competing Risk hazard estimation}
\description{
The hrcompetdx package provides a non-Parametric Diagnostic Tool for Competing Risk hazard estimation.
}
\section{competdiag functions}{

The hrcompetdx functions ...
}

