% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrcomprisk.R
\docType{package}
\name{hrcomprisk}
\alias{hrcomprisk}
\alias{hrcomprisk-package}
\title{hrcomprisk: A non-Parametric Diagnostic Tool for Competing Risk hazard estimation}
\description{
The hrcomprisk package provides a non-Parametric Diagnostic Tool for Competing Risk hazard estimation.
}
\section{competdiag functions}{

The hrcomprisk functions ...
}

