% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_CIF.R
\name{bootCRCumInc}
\alias{bootCRCumInc}
\title{Bootstrap for estimated CIF}
\usage{
bootCRCumInc(df, exit, event, exposure, entry = NULL, weights = NULL,
  ipwvars = NULL, rep = 0, seed = 54321, print.attr = T)
}
\value{
Estimating CIF per event and exposure level
}
\description{
This function is based on the CRCumInc from this XX package.
Returns an object type dataframe with the following variables and order:
col1   col2    col3  col4
ttF1.lower  ttF1.upper  ttF2.lower  ttF2.upper
}
