% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_CIF.R
\name{CRCumInc}
\alias{CRCumInc}
\title{Estimating CIF}
\usage{
CRCumInc(df, time, event, exposed, entry = NULL, weights = NULL,
  ipwvars = NULL, print.attr = T)
}
\value{
Estimating CIF per event and exposure level
}
\description{
This function is based on the CIF estimated by the survival package
Returns an object type dataframe with the following variables and order:
col1		col2		col3	col4		col5	...
event	exposure	time	CI0inc_i	CI1inc_i ...
}
