% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_ckid}
\alias{dat_ckid}
\title{CKID dataset}
\format{A data frame with XXX rows and XXX variables:
\describe{
  \item{Col1}{description of Col1}
  \item{Col2}{description of Col2}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
dat_ckid
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
