% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datcheck.R
\name{datcheck}
\alias{datcheck}
\title{Data Check for Wrapper Function}
\usage{
datcheck(df, qexit, qevent, qexposure, qentry, qweights, qipwvars,
  eoi = -1)
}
\arguments{
\item{df}{A number.}

\item{qexit}{A number.}

\item{qevent}{A number.}

\item{qexposure}{A number.}

\item{qentry}{A number.}

\item{qweights}{A number.}

\item{qipwvars}{A number.}

\item{eoi}{A number.}
}
\value{
Check dataset
}
\description{
Data needs to be of class: dataframe
Data in Long format (1 row per ID)
Package needs at LEAST colnames without any specific order:
		 "exposure" : binary (0, 1), numeric
		 "entry"	: continuous, numeric
		 "exit"		: continuous, numeric
		 "event"	: binary (>0), numeric
}
