% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_CIF.R
\name{plotCIF}
\alias{plotCIF}
\title{Plot Incidence and Ratio of sHR/csHR}
\usage{
plotCIF(cifobj, maxtime = Inf, ci = NULL, eoi = -1)
}
\value{
Plot function
}
\description{
Returns a figure
Dataset needed for function to work: the result from the data_CIF function
x is data, x default, and z is uppertime (optional) default set to 99999
this needs event, exposure, time, CI0_comp, CI1_comp,CI0inc_1,CI1inc_1, CI0inc_!1,CI1inc_!1
which is the result of CRCumInc(event)
}
